requirements.txt
absl-py==1.4.0
asgiref==3.7.2
astunparse==1.6.3
autopep8==2.0.4
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.2.0
cycler==0.12.1
Django==4.2.7
django-filter==23.3
django-silk==5.0.4
djangorestframework==3.14.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.45.0
fsspec==2023.10.0
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gprof2dot==2022.7.29
grpcio==1.54.0
h5py==3.8.0
huggingface-hub==0.19.4
idna==3.4
jax==0.4.8
Jinja2==3.1.2
joblib==1.3.2
keras==2.15.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.0
Markdown==3.5.1
MarkupSafe==2.1.2
matplotlib==3.8.2
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.3
Pillow==10.1.0
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.11.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0
termcolor==2.2.0
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.14.1
